refactor(backend): refactor TransactionalEmailAPI to ports and adapters
