+++ packages/backend/src/modules/notifications/index.ts
@@ -1,1 +1,2 @@
 export { NotificationsModule } from './NotificationsModule';
+export { TransactionalEmailAPI } from './ports/TransactionalEmailAPI';
