+++ packages/backend/src/modules/notifications/ports/TransactionalEmailAPI.ts
@@ -1,0 +1,3 @@
+export interface TransactionalEmailAPI {
+  sendMail(input: { to: string; subject: string; text: string }): Promise<void>;
+}
